mnemosine/backend/requirements.txt
# ── Core ──────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
Pillow>=10.0

# ── OpenAI Provider ──────────
openai>=1.20.0

# ── Testing ──────────────────
pytest>=7.0
httpx>=0.24

# ── HuggingFace Provider (optional — install separately for local inference) ──
# Uncomment these when using provider="hf":
# torch>=2.0
# transformers>=4.40
# accelerate>=0.30
